pyproject.toml
[project]
name = "slides-rag"
version = "0.1.0"
description = "Slides RAG"
readme = "README.md"
requires-python = "<3.13,>=3.10"
dependencies = [
    "dspy-ai>=2.5.20",
    "mistralai>=1.1.0",
    "qdrant-client>=1.12.1",
]
